This is marked __WIP__ because it's not yet fully tested, but I wanted to demonstrate what was possible due to the design of the `Utf8String` and `Utf8Span` types.

Since these types are contracted to carry only well-formed UTF-8 data, we can make certain optimizations that wouldn't be allowable in more generalized APIs like `UTF8Encoding`. For example, when computing the number of `char`s that would result from a particular `Utf8String` or `Utf8Span` instance, we can suppress all error checking and go down highly optimized code paths.

The table below shows the comparison between the baseline `Encoding.UTF8.GetChars(ReadOnlySpan<byte>)` method and the `GetCharCount` methods introduced as part of this PR. The difference is substantial, especially for code paths which weren't previously vectorized.

|            Method |      Corpus |       Mean |     Error |    StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |------------ |-----------:|----------:|----------:|------:|--------:|------:|------:|------:|----------:|
| **GetCharCount_Base** |    **11-0.txt** |  **41.112 us** | **0.7052 us** | **0.6596 us** |  **1.00** |    **0.00** |     **-** |     **-** |     **-** |         **-** |
| GetCharCount_Exp1 |    11-0.txt |  10.576 us | 0.1957 us | 0.1831 us |  0.26 |    0.01 |     - |     - |     - |         - |
|                   |             |            |           |           |       |         |       |       |       |           |
| **GetCharCount_Base** |      **11.txt** |   **3.574 us** | **0.0469 us** | **0.0439 us** |  **1.00** |    **0.00** |     **-** |     **-** |     **-** |         **-** |
| GetCharCount_Exp1 |      11.txt |   3.446 us | 0.0598 us | 0.0530 us |  0.96 |    0.02 |     - |     - |     - |         - |
|                   |             |            |           |           |       |         |       |       |       |           |
| **GetCharCount_Base** | **25249-0.txt** |  **74.032 us** | **1.3925 us** | **1.2344 us** |  **1.00** |    **0.00** |     **-** |     **-** |     **-** |         **-** |
| GetCharCount_Exp1 | 25249-0.txt |  10.977 us | 0.2190 us | 0.2769 us |  0.15 |    0.00 |     - |     - |     - |         - |
|                   |             |            |           |           |       |         |       |       |       |           |
| **GetCharCount_Base** | **30774-0.txt** |  **82.685 us** | **0.8982 us** | **0.8402 us** |  **1.00** |    **0.00** |     **-** |     **-** |     **-** |         **-** |
| GetCharCount_Exp1 | 30774-0.txt |   6.412 us | 0.1154 us | 0.2080 us |  0.08 |    0.00 |     - |     - |     - |         - |
|                   |             |            |           |           |       |         |       |       |       |           |
| **GetCharCount_Base** | **39251-0.txt** | **136.402 us** | **2.4733 us** | **2.3136 us** |  **1.00** |    **0.00** |     **-** |     **-** |     **-** |         **-** |
| GetCharCount_Exp1 | 39251-0.txt |   8.044 us | 0.1570 us | 0.1392 us |  0.06 |    0.00 |     - |     - |     - |         - |

Even small non-English texts see notable improvement. For example, pulling the word "αδαμάντινης" from one sample text and running the benchmark just on that one word results in around twice the throughput.

|            Method |      Mean |     Error |    StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |----------:|----------:|----------:|------:|------:|------:|------:|----------:|
| GetCharCount_Base | 18.665 ns | 0.3949 ns | 0.5135 ns |  1.00 |     - |     - |     - |         - |
| GetCharCount_Exp1 |  8.796 ns | 0.1680 ns | 0.1489 ns |  0.48 |     - |     - |     - |         - |

This PR only concerns itself with counting chars, not with performing the actual transcoding (creating the `string` instance) step. But it's not a giant leap to imagine having an optimized transcoder that gives `Utf8String.ToString()` markedly better performance than `Encoding.UTF8.GetString(ReadOnlySpan<byte>)`.