@GrabYourPitchforks 2019-11-07
This isn't going to be finished before the coming repo merge. I've migrated it to a temporary holding area https://github.com/GrabYourPitchforks/coreclr/pull/2 so that it doesn't get lost in the shuffle. Will reopen at some point in the future once the dust settles.